backend/sales-ai/requirements.txt
fastapi
email-validator
pydantic[email]
python-multipart
google-genai
